feat(extension): side-panel cowork bar + fix new-chat/subagent conversationId
Surface the agent's plan, workspace files, and context in the narrow
side-panel composer, and fix a conversationId resolution bug that broke
executePython and the fs tools on new chats and inside subagents.

Cowork bar (UI):
- Extract the shared cowork UI out of the home CoworkPanel into
  components/cowork/ (CoworkCard, ProgressCard, WorkingFolderCard,
  ContextCard, helpers, hooks) so home and side panel share one source;
  CoworkPanel collapses to a thin composition.
- Add CoworkBar to the chat composer: a tabbed strip (Plan / Workspace
  files / Context) with a dynamic label, click-to-expand tabs, click the
  active tab or the strip background to collapse, and a measured height
  animation on open/close/tab-switch. Hidden until there's content; panel
  content is mounted only while open so its polling hooks don't run while
  collapsed.
- Files/Context tabs render only in the side panel (showWorkspaceControls);
  the home view keeps its RightRail. In-panel card headers are suppressed
  via a new CoworkCard showHeader prop since the strip labels the panel.
- Side-panel file clicks open an in-panel FileViewerPanel (moved to
  components/files/) via FileSelectionContext.
- Animate the StepGroup "Completed N steps" fold with Collapsible.

conversationId fix:
- executePython and the fs tools (Read/Write/Edit/Glob/Grep/LS) resolve
  ctx.session?.conversationId at call time instead of from a build-time
  closure param, matching the existing delegate precedent. This fixes
  "No conversation context" on new chats (the transport was built with a
  null id and never rebuilt) and the fs tools writing to the wrong
  workspace root, and corrects subagents reusing the parent's tool
  instances with the child context.
- Drop the now-unused conversationId param from createBrowserToolSet,
  createFsTools, and createPythonTool.

Tests: add execute-python-cid and fs-cid tests covering the parent,
subagent, and null-guard paths. pnpm compile/test/build all pass (966
tests).
