+++ apps/extension/src/components/cowork/request-close-agent-tabs.ts
