+++ apps/extension/src/components/files/FileViewerPanel.tsx
